Add LLM-generated documentation
In my experience many LLMs generate decent documentation for simple code segments that are typical use scenarios. This appears to be especially true for school assignments that often cover the introductory concepts and there is a vast number of public repositories from where to get learning material for the models. Documentation is useful as it makes returning to previous code easier. While I have done the exercises myself, I see no harm in outsourcing documentation generation to LLMs as so far writing doc has not been an exercise